requirements.txt
streamlit
speedtest-cli
plotly
